oai-finetuning/pyproject.toml
[tool.poetry]
name = "vision-retrieval-oai-finetuning"
version = "0.1.0"
description = ""
authors = ["Kilian Bartz <s4kibart@uni-trier.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
jsonlines = "^4.0.0"
pillow = "^11.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
